LayoutTests/platform/ipad/imported/w3c/web-platform-tests/html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-non-integer-innerheight-expected.txt


PASS features "innerheight=/404" should NOT set "height=404"
PASS features "innerheight=_404" should NOT set "height=404"
PASS features "innerheight=L404" should NOT set "height=404"
FAIL features "innerheight=405.5" should set "height=405" assert_equals: "top=0,left=0,width=401,innerheight=405.5 value after first non-digit will be ignored" expected 405 but got 600
FAIL features "innerheight=405.32" should set "height=405" assert_equals: "top=0,left=0,width=401,innerheight=405.32 value after first non-digit will be ignored" expected 405 but got 600
FAIL features "innerheight=405LLl" should set "height=405" assert_equals: "top=0,left=0,width=401,innerheight=405LLl value after first non-digit will be ignored" expected 405 but got 600
FAIL features "innerheight=405^4" should set "height=405" assert_equals: "top=0,left=0,width=401,innerheight=405^4 value after first non-digit will be ignored" expected 405 but got 600
FAIL features "innerheight=405*3" should set "height=405" assert_equals: "top=0,left=0,width=401,innerheight=405*3 value after first non-digit will be ignored" expected 405 but got 600
FAIL features "innerheight=405/5" should set "height=405" assert_equals: "top=0,left=0,width=401,innerheight=405/5 value after first non-digit will be ignored" expected 405 but got 600
FAIL features "innerheight=405  " should set "height=405" assert_equals: "top=0,left=0,width=401,innerheight=405   value after first non-digit will be ignored" expected 405 but got 600
FAIL features "innerheight=405e1" should set "height=405" assert_equals: "top=0,left=0,width=401,innerheight=405e1 value after first non-digit will be ignored" expected 405 but got 600
FAIL features "innerheight=405e-1" should set "height=405" assert_equals: "top=0,left=0,width=401,innerheight=405e-1 value after first non-digit will be ignored" expected 405 but got 600
